moved last two checks to own file
