+++ src/ros_license_toolkit/main.py
@@ -24,12 +24,16 @@ import sys
 import timeit
 from typing import Optional, Sequence
 
-from ros_license_toolkit.checks import (LicenseTagExistsCheck,
-                                        LicenseTagIsInSpdxListCheck, Status)
-from ros_license_toolkit.license_file_referenced_check import \
+from ros_license_toolkit.checks import Status
+from ros_license_toolkit.license_checks.license_file_referenced_check import \
     LicenseFilesReferencedCheck
-from ros_license_toolkit.license_in_code_check import LicensesInCodeCheck
-from ros_license_toolkit.license_text_exists_check import \
+from ros_license_toolkit.license_checks.license_in_code_check import \
+    LicensesInCodeCheck
+from ros_license_toolkit.license_checks.license_tag_exists_check import \
+    LicenseTagExistsCheck
+from ros_license_toolkit.license_checks.license_tag_is_spdx import \
+    LicenseTagIsInSpdxListCheck
+from ros_license_toolkit.license_checks.license_text_exists_check import \
     LicenseTextExistsCheck
 from ros_license_toolkit.package import get_packages_in_path
 from ros_license_toolkit.ui_elements import (FAILURE_STR, SUCCESS_STR,
